+++ ydb/core/viewer/viewer.cpp
@@ -83,24 +83,6 @@ public:
                 .UseAuth = false,
             });
             mon->RegisterActorPage({
-                .RelPath = "viewer/json/capabilities", // temporary handling of old paths
-                .ActorSystem = ctx.ActorSystem(),
-                .ActorId = ctx.SelfID,
-                .UseAuth = false,
-            });
-            mon->RegisterActorPage({
-                .RelPath = "viewer/whoami",
-                .ActorSystem = ctx.ActorSystem(),
-                .ActorId = ctx.SelfID,
-                .UseAuth = false,
-            });
-            mon->RegisterActorPage({
-                .RelPath = "viewer/json/whoami", // temporary handling of old paths
-                .ActorSystem = ctx.ActorSystem(),
-                .ActorId = ctx.SelfID,
-                .UseAuth = false,
-            });
-            mon->RegisterActorPage({
                 .Title = "Viewer",
                 .RelPath = "viewer/v2",
                 .ActorSystem = ctx.ActorSystem(),