+++ ydb/public/lib/ydb_cli/commands/ydb_benchmark.h
@@ -26,15 +26,13 @@ private:
     TString PatchQuery(const TStringBuf& original) const;
     bool NeedRun(const TString& queryName) const;
 
-    template <typename TClient>
-    int RunBench(TClient* client, NYdbWorkload::IWorkloadQueryGenerator& workloadGen);
+    int RunBench(NYdbWorkload::IWorkloadQueryGenerator& workloadGen);
     void SavePlans(const BenchmarkUtils::TQueryBenchmarkResult& res, TStringBuf queryName, const TStringBuf name) const;
     void PrintResult(const BenchmarkUtils::TQueryBenchmarkResult& res, IOutputStream& out, const std::string& expected) const;
     BenchmarkUtils::TQueryBenchmarkSettings GetBenchmarkSettings(bool withProgress) const;
 
 private:
-    template <typename TClient> class TIterationExecution;
-    EQueryExecutor QueryExecuterType = EQueryExecutor::Generic;
+    class TIterationExecution;
     TString OutFilePath;
     ui32 IterationsCount;
     TString JsonReportFileName;